# openapi.yaml
openapi: 3.1.0
info:
  title: MOA API Contract
  description: Single Source of Truth for MOA Spring Boot 3 Backend & Flutter Frontend REST API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "POST /auth/login 응답의 accessToken을 'Authorization: Bearer <token>'으로 전달한다."
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
          example: INVALID_AUTH_TOKEN
        message:
          type: string
        details:
          type: array
          items:
            type: string
...
    AttendanceMark:
      type: string
      enum: [PRESENT, ABSENT, VACATION, UPCOMING]
      description: >-
        하루치 출석 표시 상태. PRESENT/VACATION은 attendance_records에 실제로
        저장된 값이고, ABSENT(지나간 날짜인데 기록이 없음)와 UPCOMING(오늘 중
        아직 정하지 않았거나 미래의 날짜)은 서버가 계산해서 채운다.
    MemberAttendance:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
          description: 그 인원의 User ID.
        name:
          type: string
        isMe:
          type: boolean
          description: 로그인한 사용자 본인의 행인지 여부.
        todayMark:
          $ref: '#/components/schemas/AttendanceMark'
        weeklyMarks:
          type: array
          description: 이번 주 월~일 7일치 출석 도장.
          items:
            $ref: '#/components/schemas/AttendanceMark'
          minItems: 7
          maxItems: 7
        vacationDaysUsed:
          type: integer
          description: 이번 학기 누적 휴가 사용 일수.
    AttendanceOverview:
      type: object
      properties:
        weekStart:
          type: string
          format: date
          description: 이번 주 월요일 날짜.
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberAttendance'
        myVacationDaysUsed:
          type: integer
        myVacationDaysTotal:
          type: integer
    CheckInRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: '^\d{4}$'
          description: 오늘의 출석번호 4자리.
    MyStudyInfo:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        dailyMarks:
          type: object
          description: >-
            그 달의 day-of-month(문자열 "1"~"31") -> AttendanceMark. 지나지
            않은 날(오늘 중 아직 정하지 않은 경우 포함)과 스터디가 아직
            만들어지기 전 날짜(studyCreatedAt 이전)는 둘 다 이 맵에 없다.
          additionalProperties:
            $ref: '#/components/schemas/AttendanceMark'
        presentCount:
          type: integer
        absentCount:
          type: integer
        vacationDaysUsed:
          type: integer
        vacationDaysTotal:
          type: integer
        studyCreatedAt:
          type: string
          format: date
          description: >-
            이 스터디가 만들어진 날짜. 조회한 달과 무관하게 항상 같은 값을
            돌려준다 — 프론트엔드가 월 이동 화살표/달력 선택의 하한선으로 쓴다.
paths:
...
  /clubs/{clubId}/attendance/overview:
    get:
      summary: 스터디 출석 현황 탭 조회
      description: >-
        이번 주(월~일) 전체 인원의 출석 도장과, 로그인한 사용자 본인의 이번
        학기 누적 휴가 사용/총 일수를 함께 내려준다. 가입한 스터디에 대해서만
        조회할 수 있다.
      operationId: getAttendanceOverview
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AttendanceOverview'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND)이거나 가입하지 않은 동아리(CLUB_MEMBERSHIP_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/attendance/check-in:
    post:
      summary: 출석번호로 오늘 출석 처리
      description: >-
        "출석 하기" 버튼의 출석번호 입력 제출. 오늘 발급된 번호와 일치하면
        오늘을 출석(PRESENT)으로 기록한다. 이미 기록이 있으면(예: 휴가로
        되어 있던 경우) 새로 만들지 않고 그 행을 출석으로 바꾼다.
      operationId: checkInAttendance
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '200':
          description: 출석 처리 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    nullable: true
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: >-
            존재하지 않는 동아리(CLUB_NOT_FOUND), 가입하지 않은 동아리
            (CLUB_MEMBERSHIP_NOT_FOUND), 또는 오늘의 출석번호가 아직 발급되지
            않음(ATTENDANCE_CODE_NOT_ISSUED)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '409':
          description: 출석번호가 올바르지 않음(INVALID_ATTENDANCE_CODE)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/attendance/vacation:
    post:
      summary: 오늘 휴가 사용
      description: >-
        "출석 하기" 버튼의 휴가 사용 제출. 오늘을 휴가(VACATION)로 기록한다
        (결석으로 집계되지 않음).
      operationId: useAttendanceVacation
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 휴가 처리 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    nullable: true
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND)이거나 가입하지 않은 동아리(CLUB_MEMBERSHIP_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '409':
          description: 이번 학기 휴가를 이미 모두 사용함(VACATION_LIMIT_EXCEEDED)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
  /clubs/{clubId}/attendance/me:
    get:
      summary: 내 정보 탭의 월별 출석 달력 및 통계 조회
      description: >-
        year/month로 지정한 달의 출석 달력과 이번 학기 누적 휴가/출석 통계를
        내려준다. 미래 달을 조회하면 전부 빈 상태(0)로 내려온다.
      operationId: getMyStudyInfo
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/MyStudyInfo'
                  error:
                    nullable: true
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'
        '404':
          description: 존재하지 않는 동아리(CLUB_NOT_FOUND)이거나 가입하지 않은 동아리(CLUB_MEMBERSHIP_NOT_FOUND)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    nullable: true
                  error:
                    $ref: '#/components/schemas/ApiError'